+++ tests/test.rs
@@ -1,7 +1,6 @@
 extern crate cc;
 extern crate tempdir;
 
-use std::env;
 use support::Test;
 
 mod support;
@@ -109,26 +108,6 @@ fn gnu_warnings_overridable() {
 
     test.cmd(0)
         .must_have_in_order("-Wall", "-Wno-missing-field-initializers");
-}
-
-#[test]
-fn gnu_no_warnings_if_cflags() {
-    env::set_var("CFLAGS", "-Wflag-does-not-exist");
-    let test = Test::gnu();
-    test.gcc().file("foo.c").compile("foo");
-
-    test.cmd(0).must_not_have("-Wall").must_not_have("-Wextra");
-    env::set_var("CFLAGS", "");
-}
-
-#[test]
-fn gnu_no_warnings_if_cxxflags() {
-    env::set_var("CXXFLAGS", "-Wflag-does-not-exist");
-    let test = Test::gnu();
-    test.gcc().file("foo.c").compile("foo");
-
-    test.cmd(0).must_not_have("-Wall").must_not_have("-Wextra");
-    env::set_var("CXXFLAGS", "");
 }
 
 #[test]
